Move LokaliseApi out of internal package
